requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
h11==0.9.0
httpcore==0.10.2
humanize==2.5.0
idna==2.10
mysqlclient==2.0.1
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
sniffio==1.1.0
sqlparse==0.3.1
urllib3==1.25.10
websockets==8.1
whitenoise==5.2.0
